Playwright with backend (#259)
Set up Playwright to enable anonymous users to log in, and add a false positive test to check whether login fails as we expect with a non-existing username/password pair. This means the backend server is now running during frontend UI tests with playwright.

anonymoussignup.spec.ts can be extended or used to create a util for having a per-test anonymous user available for testing logged in parts of mondey now.