Much improved: Much cleaner flow of information and code:
The router basically only acts as a thin controller, which calls the import manager. Peice by piece the import maanger validates and saves the uploaded CSV, then runs the additioanl data import (the data manager of the import manager will be set to run on that data). This way the "business logic" of importing and hardcoding lives in ImportManager, while the data, file validation, databse connection and path information lives in the data_manager